+++ CHANGELOG.md
@@ -4,10 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Fixed
-* `flexberry-groupedit` component:
-    * Fix check `hasMany` changes on route leave.
 * `flexberry-lookup` component:
     * Fix `placeholder` in dropdown mode.
+
+## [0.8.4] - 2017-02-07
+### Changed
+* Update dependency on `ember-flexberry-data` to version 0.8.2.
+
+### Fixed
+* `flexberry-groupedit` component:
+    * Fix check `hasMany` relationships changes on route leave.
 
 ## [0.8.3] - 2017-01-31
 ### Fixed